Prettied up map a bit
Experimented with the double build setup (one for the moderator, on webgl  as spectator and one as vr user), but it won work, as photonvoice doesn't allow the use of microphone or its audio in webgl;
Fixed all the problems the previous experiment brought.
Fixed animations and problems with head rotation sync with the model.
Only tracks rotation from headset.